Resolved a major performance bottleneck. Each time a FunctionCall is evaluated, the function (e.g. symbol '[') needs to be resolved: renjin needs to check each enclosing environment for a function called '[' before applying the primitive '[' function.
Previously, this involved a lot of expensive lookups in successive enclosing
HashFrames before getting to root of the environment tree. Since in 99.999%
of the cases, high-frequency functions like `if` or `+` have not been
redefined, this is just a waste.

This revision reduces this wasted time by:
- Centralizing Symbol instances in a global symbol table so that we can use
reference equality when comparing Symbols rather than equals()
- Each Symbol has an associated hash bit in the range of 1:32. High frequency
functions that are rarely redefined (like '{' or 'if') are assigned one of the
reserved bits 1, 2, or 3 depending on the likelyhood that they will redefined.
Other symbols are hashed based on their initial letter. (See SymbolTest for 
a graphic demonstration)
- HashFrame maintains a bitset that indicates which functions are present in
the frame. So when checking whether a HashFrame contains a function called
'if', we just have check the first bit, rather than doing an expensive HashMap
lookup. 

This speeds up some benchmarks by as much as 50%.




git-svn-id: https://renjin.googlecode.com/svn/trunk@338 b722d886-ae41-2056-5ac6-6db9963f4c99