Adds seperate save() and serialize() methods to RDataWriter
When R workspaces are 'saved' there are an extra 4 bytes prepended to
the stream ('RDX\n').

The serialize() functions do not add these extra bytes: there is only
a simple 'X\n' the precedes the payload.

Call save(SEXP) to add the extra 4 bytes, serialize(SEXP) for the
bare stream.



git-svn-id: https://renjin.googlecode.com/svn/trunk@588 b722d886-ae41-2056-5ac6-6db9963f4c99