Removes default library/classpath behavior (Addresses Issue #12)
Up to now, Renjin would examine the classpath and list every jar and
folder as a library location so that you could conveniently just place
packages as jars on the classpath and load them.    
                             
Unfortunately this means opening up all of these jars and doing quite
a bit of looking around at startup. I've looked at trying to speed this
up, but I don't think it will be possible with the way the R package
loading code is written, and I don't want to rewrite this code at the
moment. 

Now if you want to load a package you will need to use the lib.loc
argument of library, e.g:

library('aspect', lib.loc='src/test/resources')


git-svn-id: https://renjin.googlecode.com/svn/trunk@596 b722d886-ae41-2056-5ac6-6db9963f4c99