Initial check in of gcc-bridge for C/C++/Fortran to JVM bytecode
This is an experiment to see how far we can go with C-to-Java translation. 
Here's the idea:

GCC can be told to dump a very readable intermediate form called Gimple,
which we can in turn translate to Jimple, a very similar three address form
used by the Soot toolkit, which can be in turn compiled to JVM byte code.

This approach will surely hit some walls, but it should be enough to compile
the majority of Fortran and C code used by R packages to JVM bytecode.

git-svn-id: https://renjin.googlecode.com/svn/trunk@602 b722d886-ae41-2056-5ac6-6db9963f4c99