Further improves GCC-bridge and includes compiled stats c sources in build
So now we have a bit of functionality from the stats package compiled directly from
C to JVM bytecode that Renjin can use. With some additional work on gcc-bridge
we should be able to start leveraging a large portion of existing C/Fortran code,
porting to Java only when specific needs are identified.


git-svn-id: https://renjin.googlecode.com/svn/trunk@611 b722d886-ae41-2056-5ac6-6db9963f4c99