Introduces just-in-time compiler for vector operations
This implementation arose out of a specific pain point with dcor from the energy package,
which was blowing up in terms of memory usage and poor performance, but I think this 
will actually be the foundation for a tremendous performance win for Renjin. 

The way this works is that we put off any computation until the last possible minute by
returning a DeferredComputation, an implementation of Vector, whenever the input is 
huge. If all goes well we end up with a big nested set of vector computations to do, like
mean(sqrt(dist(a) - t(dist) - mean(a)) * sqrt(dist(b) - t(dist) - mean(b))
 
Thanks to the DeferredVector committed earlier, renjin will already do this in one loop
via virtual dispatch, but that goes pretty slowly because the JVM (at least Hotspot 6)
doesn't inline the virtual function calls, so for each of your 25m elements, there's a big
nasty chain of virtualinvokes.

To speed this up, we can employ the DeferredJitter, which takes an vector computation 
like that above and compiles it to bytecode, inlining all the accessors. This gives us a speed 
up around an order of magnitude. 

So far the jitting only works for a subset of functions, which are simply unary or binary functions
of a vector like *, /, or cos(), as well as two summary functions, rowMeans() and mean() but
I think we can leverage the asm library to transform the existing java implementations of functions
in the future.


git-svn-id: https://renjin.googlecode.com/svn/trunk@630 b722d886-ae41-2056-5ac6-6db9963f4c99