Adds special handling for JVM long int values
Even though R does not (yet) have a 64-bit integer type, we need to be able 
to at the very least accept long values from JVM code and pass them back to JVM code
without mangling them. 

To do this, we wrap any long value passed to Renjin in an LongArrayVector, which will
appear to R code as a double vector. If the vector is passed back to Java, it will be unwrapped
as a long.

git-svn-id: https://renjin.googlecode.com/svn/trunk@644 b722d886-ae41-2056-5ac6-6db9963f4c99