Google Code Issue #61: Vector pipeliner goes crazy
This is a quick fix to limit the growth of deferred computation
graphs. As submitted by Radford Neil, if you have a for loop
with 1m+ iterations that builds on the calculation of the previous
iteration, renjin will keep deferring work and end up with a monsterous
computation graph, where the costs of the book keeping FAR outweighs
the costs of doing the calculation and allocating space.

This commit adds a getComputationDepth() member of Vector that allows
us to limit the growth of the graph by periodically simplify when
the graph gets too large. This is still very slow, but at least
the computation will complete :-)