Corrected handling of transitive package dependencies
In GNU R, a package can have dependencies that are listed
in either the "Imports" or "Depends" section of the DESCRIPTION
file.

Dependencies which are listed in the "Depends" section are
actually loaded into the global search path when the primary
package is loaded with library() or require().

This effectively means that all the symbols in packages in
"Depends" sections are available to the package's namespace,
_without being explicitly imported_ in the NAMESPACE file.

This is not ideal, but since many packages depend on the
symbols in packages listed in "Depends" but not explicitly
imported, we need to support it.