Translated storage.mode<- from R to Java to improve performance
We are getting to the point where R code can be faster than
Java code in many cases where the Renjin's optimizer can defer
or optimize the R code -- while the java remains opaque and
requires hand optimization.

In this case, storage.mode<- can be rewritten in terms of
as.character(), as.double(), etc, which are primitives that will return
views over large vectors.