Force GCC to use 32-bit, regardless of platform
Previously gcc was only invoked with -m32 if Java reported
a 64-bit architecture. However, the 'os.arch' value may
be misleading under some circumstances, such as 32-bit JVM
running on a 64-bit machine.

GCC doesn't complain if -m32 is used on 32-bit platforms,
so now we always invoke GCC with this argument.