Fixed argument matching: when a function has an ellipses (...) as an argument, partial matching is NOT performed
This is a subtle difference that was leading to disappearing data.frame columns when printing to console.
print.data.frame calls format.data.frame which invokes do.call('data.frame', list.of.columns).

If any of the column names partially matched the arguments of data.frame, such as check.names, it disappeared
from the list.

Fun times.