Cleaned up number parsing code, and increased unit test coverage.
parseDouble() has been refactored to operate on a subset of a CharSequence, which allows us
to parse columns in text files without having to first create a String object for each value.

parseInt() has been removed. Upon review of GNU R's behavior and sources, it seems that actually
that integer literals in R have the same specification as floating-point literals.