Restructured compiler to generate one classfile per compilation unit
Previously we were generating one giant classfile with all functions and
global variables. This ignored the fact that not all symbols are exported
from compilation units, so two units might have functions or global
variables with the same name, such as in the survey package.

This also allows us to generate debugging information for classfiles,
enabling line-by-line debugging of C/Fortran files.