Ensured structs can be correctly referenced across compilation units.
Since each compilation unit stands alone, if two C source files share
a structure referenced in a header file, they will get different ids.

This is fine for native code generation because all that matters is
that the structs have the same size. This is a problem for targeting
the JVM, however, because the structs must also share the same
_name_ in order to be used by functions in different compilation units.

The RecordTypeDefCanonicalizer compares does a pre-pass on the
Gimple input to merge record type defs that have the same layout
across all compilation units.