Added support for compiling packages with Makevars
While 'nice' packages like the survival package simply supply
C sources, other packages have more compilicated build processes
which are encoded in a `Makevars` file which is combined with
a default `Makefile` to build a package.

I've updated the cli builder to try to reproduce the makefile step
used by GNU R, adding only the neccessary compiler flags to generate
the gimple that we need for our compilation process.