Corrected handling of POSIXlt, POSIXct conversions
In R, the fields of an POSIXlt object (year, mon, mday, hour, etc) can actually be
negative or out-of-range, for example, mday might be -4 or +3005.

This commit updates the PosixLtVector code to properly handle these types of
values, which are extensively used by seq.Date().

I've also made sure that the tzone attribute is properly interpreted and maintained
between calls of as.POSIXlt and as.POSIXct.  POSIXct values are not actually relative
to timezones, but preserving the attribute ensures that lt == as.POSIXlt(as.POSIXct(lt)),
which is important.