gcc-bridge: Fixed handling of unsigned integer casts and bitwise operations
The JVM does not intrinsically support unsigned integer types.
Fortunately, because of the way the JVM represents integers,
a form called "Two's complement", _most_ operations are the
same as they would be if the values were unsigned.

This commit fixes a few of the exceptions, including bitwise
right shift, for which the JVM offers an unsigned version, and
some casts.

Tests are pulled from the bitops package, which should now work correctly.