Started refactoring the way we generated code for C/C++ structs/objects.
We've actually been using two different strategies a bit incoherently:
sometimes representing pointers to records as normal JVM references,
sometimes as a fat pointers using ObjectPtr.

The former is definitely preferable, but can only be employed when we're
sure that the pointer references exactly one record instance:

if it points to an array of records, for example, it can't be stored as
a normal JVM reference: we need to stick in a fat pointer.

Ideally, this should be decided on a variable-by-variable basis using
an interprocedural flow analysis, but as a first step,
we can just make the decision for each distinct record type based on
a few simple rules.

Basically, if any arrays of a given record type are stored, or if there
is an expression matching the shape (record_t)malloc(size) where
size != sizeof(record_t), then we use fat pointers for this type everywhere,
otherwise, we use the more efficient unit pointer representation.