Use the provided record prefix even if the record type has a name
Multiple test cases were defining a "Rectangle" struct, which were
all given the class name "org.renjin.gcc.Rectangle".

Depending on the order of test execution, it could happen that
a Rectangle struct from an earlier test was already loaded and used
in place of the newly generated class file of the same name.

I've updated the GimpleCompiler to use the provided record prefix
so that distinct classes (e.g. cppclass$Rectangle) are generated