gcc-bridge: Implemented support for malloc() with unknown type
Usually we handle malloc's by infering the type of the result, and then
mappping the call to the equivalent JVM construct: either a new array, or
a new class.

However, in some cases, it's not possible to infer the type. Usually because
there is some a few calls between the malloc() call and the (typed) usage
of the result.

The idea is that, in these cases, malloc() will return a MallocThunk instance,
which contains only the number of bytes to allocate.

At each point where a void* pointer is cast to a concrete type, we do
a runtime check to see if the void pointer is an instance of a MallocThunk
and, if so, allocate the object of the correct size and memoize the result
with the MallocThunk.