Fixes to serialization affecting saveRDS()
saveRDS() was actually using the wrong format (that used by save() ) and so the
output couldn't be read in by GNU R.

The serialization of strings was also off on bitwise level because GNU R encodes
some strings using ASCII encoding, while we were writing all string as UTF-8.

The fix here is to write strings with code points exclusively in the area of overlap
with ASCII [0,127] marked as ASCII and all others tagged as UTF-8. In all cases,
the byte representation is still UTF-8.