gcc-bridge: Defer class cast exceptions to runtime
We just aren't set up at the moment to be able to cast a pointer
to uint32 values to a pointer to uint8, for example. But rather than
throw a compiler exception, generate the bytecode to try the cast
at runtime, which will generate an error. This may allow the code
to be partially useable.