gcc-bridge: Split Expr interface into two seperate hierarchies
Have made the split between JExpr objects, which correspond to single
JVM values and generate load/store bytecode directly, and GExpr objects,
which usually contain one or more JVM values but have semantics specific
to Gimple expressions.

For example, previously both a value of a record and a pointer to a record value
would have both been generated by SimpleExpr,  because they have the
same JVM representation - they are both stored as references to a Java class.

However, storing a value to a record *pointer* means generating a simple
ALOAD/ASTORE pair of instructions, while storing to a record *value* means
invoking the copyFrom method on the record class to copy the field values
from one record value to the other.

When these were both passed around as SimpleExprs, it got confusing and it
was easy to make errors. Now the difference between these types is probably
enforced by the (compiler's) compiler.