Fixes #230: Do not assume that we can modify "output" parameter of filter
In many cases this will work, as long as the output parameter is a StringArrayVector.
However, past a certain threshold, c() returns a lazy vector and then REAL(out) will
return a newly allocated array, changes to which will not be reflected in the original view.

Strictly speaking, this is a bug in the C code in the GNU R code, the authors rely on the
behavior of the calling R code.

With a bit more analysis, we can detect these cases and perform a defensive copy of
parameters to .Call functions, but this is a quick fix for now.