Disabled explicit NA checks for operations that themselves handle NA properly
Many of the proccessor's floating point operations themselves propogate the
extra payload associated with R's floating point NA value. So if you ask the processor
to compute 1 + NA, you'll get NA specifically, not "NaN". This means that there's no
need to branch on NA and handle it specifically.

In addition to performance considerations, the presence of explict NA checks
also affects the semantics of R operators. For example, with an explicit
NA check, NaN / NA evalutes to NaN, without the NA-specific payload.