Reduce cost of return() function calls
Normally when Exceptions are created they also spend time collecting
a stack trace for diagnostic purposes. For ControlFlowExceptions such
as ReturnException, we are not interested in the diagnostic information and
only need to use the exception to control program flow.

Java 1.7 now allows us to create an exception without initializing the stacktrace,
yielding dramatic improvements in some case.