Made gc() a NO-OP
R developers often use calls to gc() in an effort to improve the performance of
GNU R's garbage collector. Although it is possible to force the JVM to synchronously
run a round of garbage collection, it's effect on performance in Renjin is almost always
negative, as the garbage collector would otherwise be running concurrently.