Add support for active bindings (#309)
* Add new HashMap<Symbol, Closure> field to Environment for
storing active bindings

* Add to the Environment method for making active bindings
(makeActiveBinding) and checking if variable is active
(isActiveBinding).

* Change getVariable and setVariable formals to include
context. This allows getVariable and setVariable to invoke
active bindings.

* Add getVariableUnsafe and setVariableUnsafe which can't
invoke active bindings and assume that input symbol is not an
active binding.

* Change findVariable to look in active bindings, while
findVariableUnsafe is used when variable is for sure not an
active binding.

* Add support for write/read serialized R objects to/from
file.

* Make necessary changes throughout the code base to handle
active bindings such as in ls(), get(), mget(), import(),
rm(), exists(), and varArgs handeling.

* Bail in compiling loops if any of used variables is an
active bindings.

* Add depreciated functions setVariable and getVariable
without context argument.

* Add test cases for active bindings and documentation for
new and changed functions.