Change internal representation of S4 bit
Renjin previously had been using an attribute named
"__S4_BIT" to flag objects that had been "blessed" as S4 objects,
whereas GNU R stores this information in a bit within the SEXPREC.

This however led to problems because Renjin's __S4_BIT attribute
is visible in many places such as attributes(x) and in some places
was getting lost or preserved at the wrong times because it was
being treated as an attribute or a slot by the R code in the methods
package.

This commit moves the S4 bit value to an otherwise invisible field
of the AttributeMap class and updates the serialization code to be
consistent with GNU R.

However, for the moment, SEXPs written via Renjin's serialization will
*also* write out the named __S4_BIT attribute to retain compatiblity
with previous version of Renjin.

This extra attribute may be removed in the future.