Support S4 dispatch for primitive functions (#313)
Implement S4 method dispatch

Here we have added support for S4 method dispatch. In S4 method dispatch, 
the classes, super classes of the formals that constituting the signature are 
used for method dispatch, hence all signature arguments are evaluated to 
get the class information. The classes and super classes of each argument 
(and "ANY") are used to generate all the possible signatures. All existing 
methods for a given generic are stored in a environment named as 
".__T__FUNCTION:PACKAGE" according to their signature. All the 
generated signatures are looked up and the method with least distance is 
returned. For distance, the class of the argument has a distance 0, with each 
level of inheritance incrementing the distance with 1 and "ANY" distance is 
the highest distance for that argument + 1. Methods are first sorted based on 
distance, then on presence of at least 1 argument with distance 0, and at last 
the distance of first argument weights twice the distance of second, and the 
second weight twice the distance of third etc. In some cases the environment 
of returned function is populated with additional variables with meta 
information. 

The group dispatch is performed similar way. In case of there is a generic 
specific method and a group method, the generic method is returned. 
Likewise, group methods for Arith, Compare, and Logic have higher priority 
than Ops.

For special functions $ and $<- methods are looked up in Namespace of 
methods package, and S4 dispatch for @<- and @ is skipped. For S4 
objects that extend S3, the S3 part is dispatched to S3 and rewrapped in 
S4 object.

Test cases have been included to validate all these properties