Fixes #318. Refines handling of list-to-atomic conversion semantics
When converting a list to an atomic length vector, zero-length atomic vectors
are converted to NA.

NULL, which is strictly speaking, a zero-length atomic vector, is a special case
and throws an error for conversions to atomic values.

character vectors are *another* special case, in which SEXPs which would other
wise throw an error are deparsed.