Cleans up equals() and hashCode() implementations
With R, we have two different concepts of "equality".

The first you could call "mathematical" equality and is linked to the "==" operator,
in which NAs are not equal to anything else, and attributes are not considered.

The second is provided by the identical() builtin, which considers two NA elements
"identical" to each other and requires types and attributes to be exactly the same.

Previously, the Java "equals" method was a mix of the two concepts. This commit
changes the implementation of SEXP.equals() to map to the "identical" builtin.