Removed Concordion testing framework
Concordion was introduced to the project earlier to make it easier to write
well-documented unit tests.

Since then, we have taken a different direction with the hamcrest library and
building a large library of unit tests written in R, and no additional "specs" have
been added to the concordion suite.

This commit converts the corcordian test to R and removes the framework
from the project.