change as.double to as.numeric in handleS4object()
as.numeric <- is.double in R, therefor, when the first argument
to as.numeric() is a S4 object, the opName has the value "as.double".
setMethod(as.double) and setMethod(as.numeric) both create a method
in `.__T__as.numeric:base` method table. To be able to find these
created methods, opName "as.double" should be changed to "as.numeric".