remove check if file exists prior to execution of configure during building of packages
Some packages (eg XML) use the configure script to modify R script
templates (in case of XML with extension '*.R.in') with information
about the system, after which the extension is changed to '*.R'.
Checking for existence of files declared in collate section of the
DESCRIPTION file will therefor fail if any of those files are to be
generated by the configure step.