Fixed and simplified handling of provided types like "SEXP" and "DllInfo"
The handling of these kind of types had previously been mixed up with
the general handling of record types, but with the introduction of
virtual pointers,  things got confused and broken.

So we're dropping the concept of "Record unit pointers" and narrowing this
use case exclusively to the case of "provided" types where are completely
opaque to the code being compiled.

"Provided" types like SEXP* cannot be dereferenced, compiled code cannot
access their values, and they cannot be allocated by compiled code. They can *only*
passed around, and in and out of compiled code as an opaque pointer.

Insisting on these restrictions, which is perfectly suited for handling SEXPs in
compiled package code, simplifies things tremendously.