Adds local.file() function
Renjin filters all file-handling functions through a virtual file system
layer to make it easier to adapt existing code to new environments where
the file system might not be available.

However, in some cases, you really need a reference to a local file, for example,
to open with the SQLite library or for performance techniques like memory
mapping.

The local.file() function accepts a list of paths or URIs and returns a list of
absolute paths on the local filesystem.

If the URIs argument references a local file, the path is simply returned, otherwise
a temporary copy is made which will be cleaned up at the end of the process.