Implement SET_TYPEOF(x, LANGSXP)
Because Renjin maps function calls and pairlist with
different java classes, we can't change an existing pairlist
to a function call.

This adds a transformation step which looks for calls to
SET_TYPEOF(x, LANGSXP) and then traces back to the call site
where allocList creates the list, and changes it to a call to
allocLang.