#386: Additional C stdlib implementations for I/O
Includes:
* ferror
* clearerr
* rewind

And a spike for _setjmp/longjmp, that will work correctly as long as longjmp is not called.
(If longjmp is called, an unhandled exception will be raised)

This fixes one of the issues affecting the read.dbc package, but will also
help with support for all packages that do I/O in C code.