gcc-bridge: implement more support for pointer arithmetic
Sometimes C code will cast a pointer to an integer, do some math with that
integer, for example, to align it to a certain boundary, and then cast it back to a
pointer.

In order to support this, we have do some static analysis to identify which
integers are actually meant to be pointer addresses, and then keep track of which
pointer belongs to this offset so that we can be prepared to cast it back.