Transform global variables to session-scoped context
The C language code in the graphics and grDevice packages makes
extensive use of global variables to store state related to
graphics devices.

This is problematic for Renjin, as it would prevent us from running
multiple Renjin sessions in the same JVM/process.

Rather than rewriting the C code to be thread safe, this commit
introduces a new "plugin" for the Gimple compiler to automatically
transform these global variables to instance fields on a "Context"
class for each package.

Renjin ensures that packages get a fresh copy of this state for each
Renjin session, allowing a package to be used concurrently from
multiple sessions safely.