Fixes integration tests
We were using different version of jackson to test dependency
resolution, but this gets confused because gcc-bridge, and by
extension, Renjin's maven plugin, also depends on jackson.

This means that gcc-bridge's version of Jackson gets loaded on to
the classpath during tests.

Since the previous commit prioritizes the plugin's dependencies,
this screws up the test.

So instead we use different versions of JTS.