Started work on "potentially" mutating GNU R API functions
In some cases, writers of C code make assumptions
that an SEXP is not "named" and therefore assume
that an API call with modify the object in place
rather than return a modified copy.

We can detect these cases and transform the C code,
but as first step, we just warn about it.