Fixes #405: Support C++ static initialization
C++ allows you to write code like this:

    static int* INITIAL_STATE = new int(13);

To implement this, GCC will actually generate
synthetic functions which perform the allocations.

These functions are added to the .init_array section
of the object file, and are invoked by the operating
system when the ELF binary is loaded.

Gcc-bridge will now mimic this behavior by invoking these
methods in the compiled classes' static initializer.