Optimization for x^-1
Interestingly, it's not even neccessary to add our own compiler
specialization for the case of x^-1.

We can achieve the same effect by adding a branch to the Ops.pow()
function, which will be inlined by the JVM, and, if the exponent
is constant, work its way into the machine code.