Refactor JIT compiler code generation phase
We want more freedom to be able to use different strategies for
storing SEXPs on the stack. For example, for atomic vectors of known
type with constant attributes, we can store them as primitive arrays.

To make this possible, we use a similar pattern to GExpr in
gcc-bridge, so that each type of storage is responsible for converting
to another kind.