Improved/corrected propogation of variables back into environment
When compiling a loop body or other SEXP, changes to local variables
may still be visible in the environment after the compiled code
terminates.

For example:

s <- 0
for(i in 1:1000) {
   s <- s + i
}
print(s)

If we compile only the `for` expression, then the value of s
must be updated in the enclosing Environment instance so that it
is visible when print(s) is evaluated.

We had been updating all local variables at each exit path, but
this actually doesn't work for more complex control flows. For example:

for(i in 1:1000) {
  for(j in seq_along(x)) {
     s <- s + 1
  }
}

If the value of x is not known at compile time, then it is possible that
x has a length of zero and the inner loop never executes. In this case,
the values of s and j should *not* be updated in the enclosing
environment (nor _could_ we, as they have no value)

What we can do is insert UpdateEnv()
statements by doing a depth-first search of the Reverse Dominance Tree.