Improve package loading for non-cran/bioconductor packages
When building, packages will include a pointer at
META-INF/org.renjin.package/{packageName} that allows the
ClasspathPackageLoader to lookup packages on the classpath by its
unqualified name.

This also delegates resolution of unqualified package names to the
PackageLoader implementation. This allows the AetherPackageLoader
to query packages.renjin.org so that we don't have to cycle through
org.renjin.cran, org.renjin.bioconductor group ids.

This does withdraw support for loading packages in the
format '{groupId}.{packageName}', which is ambiguous: only
{groupId}:{packageName} will be accepted.