Ensure package resources can be loaded in AppEngine/Servlet contexts
Some packages (like 'httr' or 'curl') include files in their
package that they expect to be able to access at runtime.

Since Renjin assembles R packages into JAR files, we have to rely
on a virtual file system layer to map R file access operations to
access in JARs.

In sandboxed servlet environments like AppEngine, this becomes more
complex because unfettered access to the File system, where the
JAR files are located, is not permitted.

We instead provide a "fake" local file system to the R session where
the file system root is equivalent to the context root. So R.home()
should return:

    jar:file:///WEB-INF/lib/renjin-core-1.0.0-SNAPSHOT.jar!/org/renjin

And file.path(find.package("httr"), "Meta"), which might evaluate in
GNU R to:

    /home/alex/R/x86_64-pc-linux-gnu-library/3.4/httr/Meta

And on the command line Renjin as:

    jar:file:///home/alex/.m2/repository/org/renjin/cran/httr/1.3.1-b24/httr-1.3.1-b24.jar!/org/renjin/cran/httr/Meta

Must evaluate in the servlet environment as:

    jar:file:///WEB-INF/libs/httr-1.3.1-b24.jar!/org/renjin/cran/httr/Meta