Fixes #434: Ensure that graphics routines can be called from any thread
This removes a problematic thread-local cache of global variables
for the graphics and grDevices packages.

The result is that cost of reading/writing global variables from
compiled C code may be slightly higher, but it will be correct.

The alternatives would be to either initialize the thread state for
ever package on every native call, or to require the user of Renjin
to manually manage ScriptEngine-to-thread mappings.

If the overhead proves to be too high, there are better optimizations
that we can pursue, such as batch retrievals of package state within
functions, etc, that won't sacrifice correctness.