Compiling the base package (sortof)
This defines a new version for serialized packages, version 3.

Version 3 allows references to closures that have been compiled
to JVM methods.

The name of the resource file is also stored in the index, which
helps avoid the problem with lower case file systems.

We now compile the base package functions, or at least those which
compile successfully, and write those class, with source information
to class files.