Add nmax argument to unique(), duplicated(), and factor()
In GNU R (3.5.3) a positive integer value for the nmax argument to
one of these functions will in fact limit the size of the hash
table. In Renjin, this value is only used as an estimate for the
capacity of the underlying hash map. Unlike GNU R, Renjin will
therefore not give an error if nmax is too small.