Ignore out-of-range pointer access [survival]
It is theoretically possible that reading beyond the bounds
of allocated memory will cause a segfault exception, but in most
cases the read will still be within memory allocated for the process
and so the read will result in garbage, but the program will continue.

The survival package is one such offender that reads, but doesn't
use out-of-range values. This catches Java ArrayOutOfBounds exceptions
for int array accesses and returns 0xCC instead.

Fingers crossed that this doesn't tank performance.